Code cleanup for testing
